interview-challenge-v2/requirements/requirements.txt
alembic==1.8.0
fastapi==0.79.0
python-dotenv
uvicorn
psycopg2-binary
python-multipart
